added apikey and agregate settings

+++ app/routes/users.py
@@ -1,9 +1,11 @@
 from typing import List
 from fastapi import APIRouter
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import Depends, HTTPException
+from fastapi.param_functions import Header
 from sqlalchemy.orm import Session
 from ..persistence.database import SessionLocal
 from ..persistence import pdmodels, userDAO
+from ..settings import API_KEY
 
 
 router = APIRouter()
@@ -17,13 +19,17 @@ def get_db():
         db.close()
 
 @router.get('/users/', response_model=List[pdmodels.User])
-async def get_users(skip: int = 0, limit: int = 10, db: Session = Depends(get_db)):
+async def get_users(skip: int = 0, limit: int = 10, db: Session = Depends(get_db), key: str = Header(None)):
+    if key != API_KEY:
+        raise HTTPException(status_code=401, detail='Não autorizado')
     users = userDAO.get_users(db, skip, limit)
     return users
 
 
 @router.post('/users/', response_model=pdmodels.User)
-async def create_user(user: pdmodels.UserCreate, db: Session = Depends(get_db)):
+async def create_user(user: pdmodels.UserCreate, db: Session = Depends(get_db), key: str = Header(None)):
+    if key != API_KEY:
+        raise HTTPException(status_code=401, detail='Não autorizado')
     db_user = userDAO.get_user_by_unique_data(db, user)
     if db_user:
         raise HTTPException(status_code=400, detail='e-mail, login ou uuid já cadastrado')
@@ -31,7 +37,9 @@ async def create_user(user: pdmodels.UserCreate, db: Session = Depends(get_db)):
 
 
 @router.get('/users/{user_id}', response_model=pdmodels.User)
-async def get_user(user_id: int, db: Session = Depends(get_db)):
+async def get_user(user_id: int, db: Session = Depends(get_db), key: str = Header(None)):
+    if key != API_KEY:
+        raise HTTPException(status_code=401, detail='Não autorizado')
     db_user = userDAO.get_user(db, user_id)
     if db_user is None:
         raise HTTPException(status_code=404, detail='Usuário não encontrado')
@@ -39,7 +47,9 @@ async def get_user(user_id: int, db: Session = Depends(get_db)):
 
 
 @router.put('/users/{user_id}')
-async def update_user(user_id: int, user: pdmodels.UserBase, db: Session = Depends(get_db)):
+async def update_user(user_id: int, user: pdmodels.UserBase, db: Session = Depends(get_db), key: str = Header(None)):
+    if key != API_KEY:
+        raise HTTPException(status_code=401, detail='Não autorizado')
     db_user = userDAO.get_user(db, user_id)
     if db_user is None:
         raise HTTPException(status_code=404, detail='Usuário não encontrado')
@@ -47,7 +57,9 @@ async def update_user(user_id: int, user: pdmodels.UserBase, db: Session = Depen
 
 
 @router.delete('/users/{user_id}')
-async def delete_user(user_id: int, db: Session = Depends(get_db)):
+async def delete_user(user_id: int, db: Session = Depends(get_db), key: str = Header(None)):
+    if key != API_KEY:
+        raise HTTPException(status_code=401, detail='Não autorizado')
     success = userDAO.delete_user(db, user_id)
     if not success:
         raise HTTPException(status_code=404, detail='Usuário não encontrado')
